+++ lib/layrz_theme.dart
@@ -25,3 +25,4 @@ export 'src/views/views.dart';
 export 'src/grid/grid.dart';
 export 'src/table/table.dart';
 export 'src/utilities/utilities.dart';
+export 'src/map/map.dart';